Extended annotation parsing in parsers module
Previously, annotations (cells and genes files of MTX format) were only parsed to get column and
row names for the expression matrix, and even then the genes were not actually persisted with the
ETL outputs. Now, the full file is read, and output from the ETL to be loaded as two separate
metadata objects for the table in skyhook.